+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * [#201](https://github.com/xenit-eu/alfresco-docker-gradle-plugin/issues/201) - Make it possible to disable rewriting log4j.properties
 * [#227](https://github.com/xenit-eu/alfresco-docker-gradle-plugin/issues/227) - Improve message for `xyz.amp is not a valid AMP` exception
+* [#231](https://github.com/xenit-eu/alfresco-docker-gradle-plugin/issues/231) - Support for `platform()` & `enforcedPlatform()` with docker-alfresco plugin
 
 ### Fixed